requirements.txt
numpy==1.18.1
opencv-python==3.4.2.16
dlib==19.19.0
Pillow==7.0.0
sklearn==0.0
imutils==0.5.3
pandas==0.24.2
scipy==1.4.1
torch==1.4.0
torchvision==0.5.0